% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{CCM}
\alias{CCM}
\title{CCM
building Compressed component matrix}
\usage{
CCM(train_pheno, train_geno, test_geno, CCN, chunk)
}
\arguments{
\item{train_pheno}{A column is the phenotype of the training group, and the order of phenotype follows the genotype of the training group.}

\item{train_geno}{A matrix of the genotype with 0, 1, and 2 for the training group. Here, the column number equals the marker number, and the row number equals the individual number of the train group. Genotype class of the training group can be “big.matrix???.}

\item{test_geno}{A matrix of the genotype with 0, 1, and 2 for the test group. Here, the column number equals the marker number, which must match the number of markers in the training group, and the row number equals the individual number of the test group. Genotype class of the test group must match the genotype class of the training group.}

\item{CCN}{The number value is the number of building compressed components.}

\item{chunk}{If RAM is enough to run, chunk can be 1. But if the RAM is not enough for big data, users can choose an appropriate number to chop the genotype into chunks.}
}
\value{
a list of CCM of train group and test group
}
\description{
building Compressed component matrix for train group and test group.
}
