% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{CCM_sel}
\alias{CCM_sel}
\title{CCM_sel}
\usage{
CCM_sel(CCM, train_pheno)
}
\arguments{
\item{CCM}{Obtained by CCM function}

\item{train_pheno}{A column is the phenotype of the training group, and the order of phenotype follows the genotype of the training group.}
}
\value{
A list include CC sets of optimization of prediction ability and CC sets of optimization of prediction error.
}
\description{
Users will select fixed CC from CCM.
}
