% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{one_predict_pheno}
\alias{one_predict_pheno}
\title{one_predict_pheno
One step finish CCFM}
\usage{
one_predict_pheno(
  train_pheno,
  train_geno,
  test_geno,
  CCN,
  chunk,
  method,
  CV,
  num_bootstrap = 50
)
}
\arguments{
\item{train_pheno}{A column is the phenotype of the training group, and the order of phenotype follows the genotype of the training group.}

\item{train_geno}{A matrix of the genotype with 0, 1, and 2 for the training group. Here, the column number equals the marker number, and the row number equals the individual number of the train group. Genotype class of the training group can be “big.matrix???.}

\item{test_geno}{A matrix of the genotype with 0, 1, and 2 for the test group. Here, the column number equals the marker number, which must match the number of markers in the training group, and the row number equals the individual number of the test group. Genotype class of the test group must match the genotype class of the training group.}

\item{CCN}{The number value is the number of building compressed components.}

\item{chunk}{If RAM is enough to run, chunk can be 1. But if the RAM is not enough for big data, users can choose an appropriate number to chop the genotype into chunks.}

\item{method}{The option can be LM, SVM, SVM_bag, LM_Max, SVM_Max, LM_Min, SVM_Min.}

\item{CV}{If OFF, the argument of method can be LM, SVM, SVM_bag, and If NO and CCN>10 , the argument of method can be LM_Max, SVM_Max, LM_Min, SVM_Min.}

\item{num_bootstrap}{It's the number of samples in bagging_SVM. The default option is 50.}
}
\value{
Predition_Pheno It is the prediction phenotype of the test group, which order follows the genotype of the test group.
}
\description{
Users will only get the prediction phenotype of test group.
}
