% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{pre_test}
\alias{pre_test}
\title{pre_test
One step finish CCFM}
\usage{
pre_test(train_pheno, train_geno, pre_test_num)
}
\arguments{
\item{train_pheno}{A column is the phenotype of the training group, and the order of phenotype follows the genotype of the training group.}

\item{train_geno}{A matrix of the genotype with 0, 1, and 2 for the training group. Here, the column number equals the marker number, and the row number equals the individual number of the train group. Genotype class of the training group can be “big.matrix???.}

\item{pre_test_num}{It's the number of samples in pre-experiment.}
}
\value{
The number of times the method with the highest prediction accuracy appeared in the pre-experiment.
}
\description{
Users will only get the prediction phenotype of test group.
}
