% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_pheno}
\alias{predict_pheno}
\title{predict_pheno}
\usage{
predict_pheno(
  method,
  CCM,
  train_pheno,
  test_geno,
  train_geno,
  num_bootstrap = 50
)
}
\arguments{
\item{method}{The option can be LM, SVM, SVM_bag, LM_Max, SVM_Max, LM_Min, SVM_Min.}

\item{CCM}{Obtained by CCM function}

\item{train_pheno}{A column is the phenotype of the training group, and the order of phenotype follows the genotype of the training group.}

\item{test_geno}{A matrix of the genotype with 0, 1, and 2 for the test group. Here, the column number equals the marker number, which must match the number of markers in the training group, and the row number equals the individual number of the test group. Genotype class of the test group must match the genotype class of the training group.}

\item{train_geno}{A matrix of the genotype with 0, 1, and 2 for the training group. Here, the column number equals the marker number, and the row number equals the individual number of the train group. Genotype class of the training group can be “big.matrix???.}

\item{num_bootstrap}{It's the number of samples in bagging_SVM. The default option is 50.}
}
\value{
Predition_Pheno It is the prediction phenotype of the test group, which order follows the genotype of the test group.
}
\description{
Users will only get the prediction phenotype of test group.
}
